app/src/frontend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
requests==2.32.3
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-logging==0.47b0
